Move all metadata stores to zenml.core.metadata subpackage level
