+++ zenml/core/metadata/metadata_wrapper_factory.py
@@ -13,9 +13,9 @@
 #  permissions and limitations under the License.
 """Factory to register metadata_wrapper classes to metadata_wrappers"""
 
-from zenml.core.metadata.mock_metadata_wrapper import MockMetadataStore
-from zenml.core.metadata.mysql_metadata_wrapper import MySQLMetadataStore
-from zenml.core.metadata.sqlite_metadata_wrapper import SQLiteMetadataStore
+from zenml.core.metadata import MockMetadataStore
+from zenml.core.metadata import MySQLMetadataStore
+from zenml.core.metadata import SQLiteMetadataStore
 
 
 class MetadataWrapperFactory:
