+++ zenml/core/metadata/metadata_wrapper_test.py
@@ -16,7 +16,7 @@ import random
 
 import pytest
 
-from zenml.core.metadata.metadata_wrapper import ZenMLMetadataStore
+from zenml.core.metadata import ZenMLMetadataStore
 from zenml.core.standards.standard_keys import MLMetadataKeys
 from zenml.utils.enums import GDPComponent
 from zenml.utils.enums import PipelineStatusTypes
